+++ bika/lims/upgrade/v01_02_003.py
@@ -4,7 +4,9 @@
 #
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
-from bika.lims import logger
+from bika.lims import logger, api
+from bika.lims.catalog.analysisrequest_catalog import \
+    CATALOG_ANALYSIS_REQUEST_LISTING
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
@@ -28,6 +30,26 @@ def upgrade(tool):
 
     # -------- ADD YOUR STUFF HERE --------
 
+    # Unbound the worksheetanalysis_workflow from Analysis Requests and add a
+    # FieldIndex 'assigned_state' in AR's catalog (for its use on searches)
+    fix_assign_analysis_requests(portal, ut)
+
     logger.info("{0} upgraded to version {1}".format(product, version))
 
     return True
+
+
+def fix_assign_analysis_requests(portal, ut):
+    # Remove 'bika_worksheet_analysis_workflow' from AnalysisRequest
+    wfid = 'bika_worksheetanalysis_workflow'
+    wtool = api.get_tool('portal_workflow')
+    chain = wtool.getChainFor('AnalysisRequest')
+    if wfid in chain:
+        # Remove the workflow from AR
+        chain = [ch for ch in chain if ch != wfid]
+        wtool.setChainForPortalTypes(['AnalysisRequest', ], chain)
+
+    # Add the `assigned_state` index for Analysis Request catalog
+    ut.addIndexAndColumn(CATALOG_ANALYSIS_REQUEST_LISTING, 'assigned_state',
+                'FieldIndex')
+    ut.refreshCatalogs()